apps/graph_writer/requirements.txt
neo4j==5.21.0
kafka-python-ng[crc32c,lz4,snappy,zstd]==2.2.3
pydantic==2.7.4
backoff==2.2.1
